fix: 구단 소개페이지에서 애니메이션 공통 컴포넌트로 추출
