+++ src/pages/Introduction/IntroductionClub.tsx
@@ -1,12 +1,17 @@
 import { IntroductionClubData } from '@/mocks/introduction/MockIntroductionClub';
-
+import IntroductionContent from '@/components/common/IntroductionContent';
 import IntroductionClubItem from '@/components/introduction/IntroductionClubItem';
 
 const IntroductionClub = () => {
   return (
-    <div className="flex flex-col gap-36">
+    <div className="space-y-36">
       {IntroductionClubData.map((data, index) => (
-        <IntroductionClubItem key={data.id} data={data} index={index} />
+        <IntroductionContent
+          key={data.id}
+          data={data}
+          index={index}
+          children={<IntroductionClubItem data={data} />}
+        />
       ))}
     </div>
   );
